# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "description": "API endpoints for evaluating risk",
    "version": "1.0.0"
  },
  "paths": {
    "/evaluar_riesgo": {
      "post": {
        "summary": "Evaluate risk based on blood data",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "paciente": {
                    "type": "string"
                  },
                  "azucar": {
                    "type": "number"
                  },
                  "grasa": {
                    "type": "number"
                  },
                  "oxigeno": {
                    "type": "number"
                  }
                },
                "required": ["paciente", "azucar", "grasa", "oxigeno"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "riesgo": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/sangre_evaluada/{paciente}": {
      "get": {
        "summary": "Get evaluated blood information by patient name",
        "parameters": [
          {
            "name": "paciente",
            "in": "path",
            "description": "Name of the patient",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "paciente": {
                      "type": "string"
                    },
                    "azucar": {
                      "type": "number"
                    },
                    "grasa": {
                      "type": "number"
                    },
                    "oxigeno": {
                      "type": "number"
                    },
                    "riesgo": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update evaluated blood information by patient name",
        "parameters": [
          {
            "name": "paciente",
            "in": "path",
            "description": "Name of the patient",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "azucar": {
                    "type": "number"
                  },
                  "grasa": {
                    "type": "number"
                  },
                  "oxigeno": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete evaluated blood information by patient name",
        "parameters": [
          {
            "name": "paciente",
            "in": "path",
            "description": "Name of the patient",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/paciente/sangre_evaluada": {
      "get": {
        "summary": "Get all evaluated blood records",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "registros": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "paciente": {
                            "type": "string"
                          },
                          "azucar": {
                            "type": "number"
                          },
                          "grasa": {
                            "type": "number"
                          },
                          "oxigeno": {
                            "type": "number"
                          },
                          "riesgo": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}